fix(): remove scripts folder + remove fluent components
