+++ src/pages/app-about.ts
@@ -1,0 +1,39 @@
+import { LitElement, css, html } from 'lit';
+import { customElement } from 'lit/decorators.js';
+
+@customElement('app-about')
+export class AppAbout extends LitElement {
+  static get styles() {
+    return css`
+
+
+    `;
+  }
+
+  constructor() {
+    super();
+  }
+
+  render() {
+    return html`
+      <app-header ?enableBack="${true}"></app-header>
+
+      <div>
+        <h2>About Page</h2>
+
+        <div>
+          <h2>Did you know?</h2>
+
+          <p>PWAs have access to many useful APIs in modern browsers! These
+            APIs have enabled many new types of apps that can be built as PWAs, such as advanced graphics editing apps, games,
+            apps that use machine learning and more!
+          </p>
+
+          <p>Check out <a
+              href="https://docs.microsoft.com/en-us/microsoft-edge/progressive-web-apps-chromium/how-to/handle-files">these
+              docs</a> to learn more about the advanced features that you can use in your PWA</p>
+        </div>
+      </div>
+    `;
+  }
+}
