--- src/script/pages/app-about.ts
@@ -1,58 +1,0 @@
-import { LitElement, css, html } from 'lit';
-import { customElement } from 'lit/decorators.js';
-
-@customElement('app-about')
-export class AppAbout extends LitElement {
-  static get styles() {
-    return css`
-
-    fluent-card {
-      padding: 0px 18px 18px;
-    }
-
-    @media(prefers-color-scheme: light) {
-      fluent-card {
-        --fill-color: #edebe9;
-      }
-    }
-
-    @media(prefers-color-scheme: dark) {
-      fluent-card {
-        --fill-color: #4e4e4e;
-        color: white;
-        border: none;
-      }
-    }
-
-    @media (min-width: 1024px) {
-      fluent-card {
-        width: 54%;
-      }
-    }
-    `;
-  }
-
-  constructor() {
-    super();
-  }
-
-  render() {
-    return html`
-      <app-header ?enableBack="${true}"></app-header>
-
-      <div>
-        <h2>About Page</h2>
-
-        <fluent-card>
-          <h2>Did you know?</h2>
-
-          <p>PWAs have access to many useful APIs in modern browsers! These
-            APIs have enabled many new types of apps that can be built as PWAs, such as advanced graphics editing apps, games, apps that use machine learning and more!
-          </p>
-
-          <p>Check out <fluent-anchor href="https://docs.microsoft.com/en-us/microsoft-edge/progressive-web-apps-chromium/how-to/handle-files">these docs</fluent-anchor> to learn more about the advanced features that you can use in your PWA</p>
-        </fluent-card>
-      </div>
-    `;
-  }
-}
